Final Update:
- Deleting some scripts as I don't need anymore.
- Writing comments in code to explain how it works.
- Making some changes in scripts.
- Making some changes in Level2 Scene.
