--- Assets/BarrierBlocked.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb3dbc6e8a5606a4598f08e1e3973714
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
