--- Assets/GameManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6707dc6f57486254084951a7d08e21ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
