--- Assets/GameOverScreen.cs
@@ -1,18 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class GameOverScreen : MonoBehaviour
-{
-    // Start is called before the first frame update
-    public void PlayGame()
-    {
-        SceneManager.LoadScene("Level2Scene");
-    }
-
-    public void ExitGame()
-    {
-        SceneManager.LoadScene("MainMenuScene");
-    }
-}
