+++ Assets/Level2SceneScripts/CheckpointSpawn.cs
@@ -5,14 +5,19 @@ using UnityEngine;
 public class CheckpointSpawn : MonoBehaviour
 {
 
-    PlayerLose player;
+    // Variables for functions below.
+
+    // Need to make a ref for the tag by having GetComponent function.
+    TwoDCharacter_Movement player;
+
+    // For the position of the checkpoint spawn after triggered it.
     public Transform checkpointspawn;
 
     // Start is called before the first frame update
     private void Awake()
     {
-        player = GameObject.FindGameObjectWithTag("Player").GetComponent<PlayerLose>();
-        //player = GameObject.FindGameObjectWithTag("Checkpoint").GetComponent<TwoDCharacter_Movement>();
+        // To tag the Player.
+        player = GameObject.FindGameObjectWithTag("Player").GetComponent<TwoDCharacter_Movement>(); 
     }
 
     // Update is called once per frame
@@ -21,10 +26,10 @@ public class CheckpointSpawn : MonoBehaviour
 
     }
 
+    // Tagging player for the Checkpoint Gameobject.
     private void OnTriggerEnter2D(Collider2D other)
     {
         if (other.CompareTag("Player"))
-        //if (other.CompareTag("Checkpoint"))
         {
             //player.UpdateCheckpoint(transform.position);
             player.UpdateCheckpoint(checkpointspawn.position);
