+++ Assets/Level2SceneScripts/CheckpointSpawn.cs.meta
