+++ Assets/Level2SceneScripts/Pickup Item.cs.meta
