--- Assets/PlayerLives.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4a4e0f9e3f769a7419bcf72d070d4e47
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
