--- Assets/PlayerLose.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3232a86c811360841bfb3fdf7e5e5c7e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
